test/requirements.txt
requests
python-socketio[client]

flake8
